[SECURITY-868] Fold tests back into normal files
